refactor: 폴더 네이밍 변경
