+++ front/src/pages/CrewMain/styles.ts
